requirements.txt
# --- Core Frameworks ---
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3

# --- LangChain & AI ---
langchain>=0.1.0
langchain-openai>=0.0.8

# --- Utilities & Environment ---
python-dotenv==1.0.1

# --- Testing ---
pytest==8.0.2